+++ src/main/java/be/baur/sdt/statements/ForEachStatement.java
@@ -85,7 +85,8 @@ public class ForEachStatement extends XPathStatement {
 	@Override
 	public DataNode toSDA() {
 		DataNode node = new DataNode(Statements.FOREACH.tag, getExpression());
-		for (Node statement : nodes()) // add child statements
+		node.add(null); // render compound statement, even if empty
+		for (Node statement : nodes()) // add any child statements
 			node.add(((Statement) statement).toSDA());
 		return node;
 	}