+++ src/main/java/be/baur/sdt/statements/IfStatement.java
@@ -61,7 +61,8 @@ public class IfStatement extends XPathStatement {
 	@Override
 	public DataNode toSDA() {
 		DataNode node = new DataNode(Statements.IF.tag, getExpression()); 
-		for (Node statement : nodes()) // add child statements
+		node.add(null); // render compound statement, even if empty
+		for (Node statement : nodes()) // add any child statements
 			node.add(((Statement) statement).toSDA());
 		return node;
 	}