+++ src/main/java/be/baur/sdt/statements/OtherwiseStatement.java
@@ -29,7 +29,8 @@ public class OtherwiseStatement extends Statement {
 	@Override
 	public DataNode toSDA() {
 		DataNode node = new DataNode(Statements.OTHERWISE.tag); 
-		for (Node statement : nodes()) // add child statements
+		node.add(null); // render compound statement, even if empty
+		for (Node statement : nodes()) // add any child statements
 			node.add(((Statement) statement).toSDA());
 		return node;
 	}