+++ src/main/java/be/baur/sdt/xpath/function/CompareNumberFunction.java
@@ -16,15 +16,15 @@ import org.jaxen.function.NumberFunction;
  * numbers and returns -1 if the second argument precedes the first, 1 if it
  * exceeds it, and 0 if the arguments are numerically equal:
  * <p>
- * <code>sdt:compare-number(1,2)</code> returns <code>-1</code>.<br>
- * <code>sdt:compare-number(3,'3')</code> returns <code>0</code>.<br>
- * <code>sdt:compare-number('5','4')</code> returns <code>1</code>.
+ * <code>sdt:compare-number(1,2)</code> returns <code>-1.0</code>.<br>
+ * <code>sdt:compare-number(3,'3')</code> returns <code>0.0</code>.<br>
+ * <code>sdt:compare-number('5','4')</code> returns <code>1.0</code>.
  * <p>
  * Objects that are not numbers are considered equal, and greater than all other
  * numbers:
  * <p>
- * <code>sdt:compare-number('a',1)</code> returns <code>1</code>.<br>
- * <code>sdt:compare-number('a','b')</code> returns <code>0</code>.
+ * <code>sdt:compare-number('a',1)</code> returns <code>1.0</code>.<br>
+ * <code>sdt:compare-number('a','b')</code> returns <code>0.0</code>.
  * <p>
  * If the optional third argument evaluates to true, objects that are not
  * numbers are considered smaller than all numbers.
@@ -39,8 +39,9 @@ public class CompareNumberFunction implements Function
 
     
 	/**
-	 * Numerically compares two values, returning -1, 0 or 1 depending on the
-	 * outcome.
+	 * Numerically compares two arguments, returning -1, 0 or 1. An optional third
+	 * argument - after boolean evaluation - determines whether NaN is considered
+	 * smaller (if true) or greater (the default) than all other numbers.
 	 *
 	 * @param context the context at the point in the expression when the function
 	 *                is called
@@ -51,33 +52,50 @@ public class CompareNumberFunction implements Function
 	 * @throws FunctionCallException if <code>args</code> has more than three or
 	 *                               less than two items.
 	 */
-    @Override
+	@Override
 	@SuppressWarnings("rawtypes")
-	public Object call(Context context, List args) throws FunctionCallException
-    {
-    	final int argc = args.size();
-        if (argc < 2 || argc > 3)
-            throw new FunctionCallException( "compare-number() requires two or three arguments." );
+	public Object call(Context context, List args) throws FunctionCallException {
 
-        final Navigator nav = context.getNavigator();
-        
-        Double d1 = NumberFunction.evaluate(args.get(0), nav);
-        Double d2 = NumberFunction.evaluate(args.get(1), nav);
-        
+		final int argc = args.size();
+		if (argc < 2 || argc > 3)
+			throw new FunctionCallException("compare-number() requires two or three arguments.");
+
+		final Navigator nav = context.getNavigator();
+
+		return evaluate(args.get(0), args.get(1), argc == 3 && BooleanFunction.evaluate(args.get(2), nav), nav);
+	}
+    
+
+	/**
+	 * Numerically compares two objects, returning -1, 0 or 1.
+	 *
+	 * @param obj1     the first object to be compared
+	 * @param obj2     the second object to be compared
+	 * @param nanFirst whether NaN is considered smaller than all other numbers.
+	 * 
+	 * @return a <code>Double</code>
+	 */
+	public static Double evaluate(Object obj1, Object obj2, boolean nanFirst, Navigator nav) {
+
+		final Double d1 = NumberFunction.evaluate(obj1, nav);
+		final Double d2 = NumberFunction.evaluate(obj2, nav);
+
 		/*
 		 * The standard Double.comparedTo() considers NaN greater than all other
 		 * doubles. So when sorting ascending, they will end up last. If <nanFirst> is
 		 * true we treat NaN as smaller than all other doubles, so they end up first.
 		 */
-        if (argc == 3 && BooleanFunction.evaluate(args.get(2), nav)) {
-        
-        	boolean nan1 = d1.isNaN(), nan2 = d2.isNaN();
-        	if (nan1 && nan2) return 0; // both objects are NaN and considered equal
-        	if (nan1 || nan2) return nan1 ? -1 : 1; // if first is NaN it is smaller, otherwise greater
-        }
-        
-        // If we get here a regular compareTo applies. Must return a double!
-        return Math.signum((double)d1.compareTo(d2)); 
-    }
+		if (nanFirst) {
+
+			boolean nan1 = d1.isNaN(), nan2 = d2.isNaN();
+			if (nan1 && nan2)
+				return 0.0; // both objects are NaN and therefore considered equal
+			if (nan1 || nan2)
+				return nan1 ? -1.0 : 1.0; // if first is NaN it's smaller, else greater
+		}
+
+		// If we get here a regular compareTo applies. Must return a double!
+		return Math.signum((double) d1.compareTo(d2));
+	}
 
 }