+++ src/test/java/test/TestSDAXPath.java
@@ -160,15 +160,15 @@ public class TestSDAXPath {
 		t.so("S86", "sdt:right(/addressbook/contact[2]/firstname,2)", doc, "ob");
 		t.so("S87", "sdt:right(/addressbook/contact[2]/firstname,4)", doc, "Bob");
 		
-		t.so("S88", "sdt:compare-number(1,2)", doc, "-1");
-		t.so("S89", "sdt:compare-number(3,'3')", doc, "0");
-		t.so("S90", "sdt:compare-number('5','4')", doc, "1");
-		t.so("S91", "sdt:compare-number('a',1)", doc, "1");
-		t.so("S92", "sdt:compare-number('a','b')", doc, "0");
-		t.so("S91", "sdt:compare-number('a',1,true())", doc, "-1");
-		t.so("S91", "sdt:compare-number('a',1,false())", doc, "1");
-		t.so("S92", "sdt:compare-number('a','b',true())", doc, "0");
-		t.so("S92", "sdt:compare-number('a','b',false())", doc, "0");
+		t.so("S88", "sdt:compare-number(1,2)", doc, "-1.0");
+		t.so("S89", "sdt:compare-number(3,'3')", doc, "0.0");
+		t.so("S90", "sdt:compare-number('5','4')", doc, "1.0");
+		t.so("S91", "sdt:compare-number('a',1)", doc, "1.0");
+		t.so("S92", "sdt:compare-number('a','b')", doc, "0.0");
+		t.so("S91", "sdt:compare-number('a',1,true())", doc, "-1.0");
+		t.so("S91", "sdt:compare-number('a',1,false())", doc, "1.0");
+		t.so("S92", "sdt:compare-number('a','b',true())", doc, "0.0");
+		t.so("S92", "sdt:compare-number('a','b',false())", doc, "0.0");
 	}
 
 }